refactor(ivy): remame `styling_next` directory to `styling`
