+++ packages/core/src/render3/instructions/element.ts
@@ -17,9 +17,9 @@ import {BINDING_INDEX, HEADER_OFFSET, LView, RENDERER, TVIEW, T_HOST} from '../i
 import {assertNodeType} from '../node_assert';
 import {appendChild} from '../node_manipulation';
 import {decreaseElementDepthCount, getElementDepthCount, getIsParent, getLView, getNamespace, getPreviousOrParentTNode, getSelectedIndex, increaseElementDepthCount, setIsNotParent, setPreviousOrParentTNode} from '../state';
-import {registerInitialStylingOnTNode} from '../styling_next/instructions';
-import {StylingMapArray, TStylingContext} from '../styling_next/interfaces';
-import {getInitialStylingValue, hasClassInput, hasStyleInput} from '../styling_next/util';
+import {registerInitialStylingOnTNode} from '../styling/instructions';
+import {StylingMapArray, TStylingContext} from '../styling/interfaces';
+import {getInitialStylingValue, hasClassInput, hasStyleInput} from '../styling/util';
 import {setUpAttributes} from '../util/attrs_utils';
 import {getNativeByTNode, getTNode} from '../util/view_utils';
