+++ packages/core/src/render3/styling/map_based_bindings.ts
