+++ packages/core/src/render3/styling/state.ts
