+++ packages/core/test/render3/perf/style_and_class_bindings/index.ts
@@ -11,7 +11,7 @@ import {refreshView} from '../../../../src/render3/instructions/shared';
 import {RenderFlags} from '../../../../src/render3/interfaces/definition';
 import {AttributeMarker} from '../../../../src/render3/interfaces/node';
 import {TVIEW} from '../../../../src/render3/interfaces/view';
-import {ɵɵclassProp, ɵɵstyleProp} from '../../../../src/render3/styling_next/instructions';
+import {ɵɵclassProp, ɵɵstyleProp} from '../../../../src/render3/styling/instructions';
 import {createBenchmark} from '../micro_bench';
 import {setupRootViewWithEmbeddedViews} from '../setup';
